+++ app/src/main/java/com/project/lab1/feed/adapters/Adapter.kt
@@ -2,14 +2,13 @@ package com.project.lab1.feed.adapters
 
 
 import android.content.Intent
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.project.lab1.databinding.FeedCellLayoutBinding
 //import com.project.lab1.databinding.ActivitySeeNotesBinding
-import com.project.lab1.feed.ImagePost
+import com.project.lab1.feedImage.ImageActivity
 import com.project.lab1.feed.models.FeedItem
 import com.project.lab1.feed.viewHolders.FeedItemViewHolder
 import kotlinx.android.synthetic.main.feed_cell_layout.view.*
@@ -35,7 +34,7 @@ class FeedRecyclerViewAdapter(private val dataSet: Array<FeedItem>) : RecyclerVi
 
     fun onClick(view: View) {
         val id = view.headerTextView.text
-        val intent = Intent(view.context, ImagePost::class.java)
+        val intent = Intent(view.context, ImageActivity::class.java)
         intent.putExtra("id", id);
         view.context.startActivity(intent)
     }
