+++ app/src/main/java/com/project/lab1/notes/adapters/NoteAdapter.kt
@@ -1,11 +1,11 @@
-package com.project.lab1.feed.adapters
+package com.project.lab1.notes.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.project.lab1.databinding.NoteCellLayoutBinding
-import com.project.lab1.feed.models.NoteItem
-import com.project.lab1.feed.viewHolders.NoteViewHolder
+import com.project.lab1.notes.models.NoteItem
+import com.project.lab1.notes.viewHolders.NoteViewHolder
 
 
 class NoteAdapter(private val dataSet: Array<NoteItem>) : RecyclerView.Adapter<NoteViewHolder>() {
