+++ app/src/main/java/com/project/lab1/notes/viewHolders/NoteViewHolder.kt
@@ -1,4 +1,4 @@
-package com.project.lab1.feed.viewHolders
+package com.project.lab1.notes.viewHolders
 
 import androidx.recyclerview.widget.RecyclerView
 import com.project.lab1.R
@@ -6,7 +6,7 @@ import com.project.lab1.R
 //import com.project.lab1.databinding.FeedCellLayoutBinding
 import com.project.lab1.databinding.NoteCellLayoutBinding
 //import com.project.lab1.feed.models.FeedItem
-import com.project.lab1.feed.models.NoteItem
+import com.project.lab1.notes.models.NoteItem
 import com.squareup.picasso.Picasso
 
 class NoteViewHolder(private val binding: NoteCellLayoutBinding): RecyclerView.ViewHolder(binding.root) {
