+++ app/src/main/res/layout/activity_see_notes.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".feed.SeeNotes">
+    tools:context=".notes.NotesActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/note_recycler_view"
